Fix signal install on unsupport arch's

+++ install/desktop/app-signal.sh
@@ -6,14 +6,16 @@ case $ARCH in
     ;;
   *)
     echo "Unsupported architecture: $ARCH"
-    return 0
     ;;
 esac
 
-wget -qO- https://updates.signal.org/desktop/apt/keys.asc | gpg --dearmor >signal-desktop-keyring.gpg
-cat signal-desktop-keyring.gpg | sudo tee /usr/share/keyrings/signal-desktop-keyring.gpg >/dev/null
-echo "deb [arch=${SIGNAL_ARCH} signed-by=/usr/share/keyrings/signal-desktop-keyring.gpg] https://updates.signal.org/desktop/apt xenial main" |
-	sudo tee /etc/apt/sources.list.d/signal-xenial.list
-rm signal-desktop-keyring.gpg
-sudo apt update
-sudo apt install -y signal-desktop
+# Only continue if architecture is supported
+if [ -n "$SIGNAL_ARCH" ]; then
+  wget -qO- https://updates.signal.org/desktop/apt/keys.asc | gpg --dearmor >signal-desktop-keyring.gpg
+  cat signal-desktop-keyring.gpg | sudo tee /usr/share/keyrings/signal-desktop-keyring.gpg >/dev/null
+  echo "deb [arch=${SIGNAL_ARCH} signed-by=/usr/share/keyrings/signal-desktop-keyring.gpg] https://updates.signal.org/desktop/apt xenial main" |
+    sudo tee /etc/apt/sources.list.d/signal-xenial.list
+  rm signal-desktop-keyring.gpg
+  sudo apt update
+  sudo apt install -y signal-desktop
+fi
